requirements.txt
# Core backend/UI
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.36.0

# PDF & text
pymupdf==1.24.5           # PyMuPDF
pdfplumber==0.11.0

# RAG
faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4
scikit-learn==1.5.1
rapidfuzz==3.9.5

# Reranker (CrossEncoder)
# Lightweight and CPU friendly

# Utilities
pandas==2.2.2